Confirmation dialog to delete an UI configuration

+++ web-ui/src/main/resources/catalog/js/admin/SystemSettingsController.js
@@ -278,6 +278,10 @@
       };
 
       $scope.deleteUiConfig = function() {
+        $('#gn-confirm-remove-ui').modal('show');
+      };
+
+      $scope.confirmDeleteUiConfig = function() {
         $scope.lastUiConfiguration = undefined;
         $http.delete('../api/ui/' + $scope.uiConfiguration.id).then(function(r) {
           loadUiConfigurations();
